+++ src/app/components/sidebar/index.jsx
@@ -2,11 +2,14 @@
 
 import React from 'react';
 import PropTypes from 'prop-types';
+import HideableComponent from "../generic/hideable";
 
 class Sidebar extends React.Component {
 
 	render() {
 
+    const { preferences } = this;
+    
 		const sections = this.sections.map(section => {
 
 			const isActive = this.activeSection === section.id;
@@ -23,13 +26,15 @@ class Sidebar extends React.Component {
 			};
 
 			return (
-				<li key={ section.id } className={ className } role="tab" id={ `tab-${section.id}` }
-					aria-selected={ isActive } aria-controls={ `tabpanel-${section.id}` } tabIndex={ isActive ? 0 : -1 }
-					aria-label={ section.label }
-					onClick={ this.selectSection.bind(this, section.id) }>
-					<div className="section-icon" style={ style } />
-					<span className="section-label">{ section.label }</span>
-				</li>
+        <HideableComponent allPreferences={ preferences } field={ section }>
+          <li key={ section.id } className={ className } role="tab" id={ `tab-${section.id}` }
+            aria-selected={ isActive } aria-controls={ `tabpanel-${section.id}` } tabIndex={ isActive ? 0 : -1 }
+            aria-label={ section.label }
+            onClick={ this.selectSection.bind(this, section.id) }>
+            <div className="section-icon" style={ style } />
+            <span className="section-label">{ section.label }</span>
+          </li>
+        </HideableComponent>
 			);
 
 		});
@@ -42,6 +47,12 @@ class Sidebar extends React.Component {
 
 	}
 
+  get preferences() {
+    
+    return this.props.preferences;
+    
+  }
+  
 	get sections() {
 
 		return this.props.sections;
@@ -130,6 +141,7 @@ Sidebar.propTypes = {
 	onSelectSection: PropTypes.func,
 	selectSection: PropTypes.func,
 	onTablistKeyDown: PropTypes.func,
+  preferences: PropTypes.object,
 };
 
 export default Sidebar;