+++ example/preload.js
@@ -8,16 +8,8 @@ const { ipcRenderer } = electron;
 let onPreferencesChangedHandler = () => {};
 
 contextBridge.exposeInMainWorld('api', {
-	showPreferences() {
-
-		ipcRenderer.send('showPreferences');
-
-	},
-	closePreferences() {
-
-		ipcRenderer.send('closePreferences');
-
-	},
+	showPreferences: (section) => ipcRenderer.send('showPreferences', section),
+	closePreferences: () => ipcRenderer.send('closePreferences'),
 	getPreferences: () => ipcRenderer.sendSync('getPreferences'),
 	onPreferencesChanged(handler) {