+++ CHANGELOG.md
@@ -5,6 +5,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- `hideFunction` on sections, groups or fields which you can conditionally hide (via other preferences) (#160)
+- Ability to pass JS functions in the preference object to the renderer (#160)
+
+### Changed
+- Bump dependencies to their latest version
 
 ## [2.7.0] - 2022-07-18
 ### Fixed